+++ zcash_client_sqlite/src/wallet.rs
@@ -5925,6 +5925,8 @@ pub(crate) fn query_transparent_spend_locator_map(
             index,
             vec![],
             vec![],
+            #[cfg(feature = "transparent-inputs")]
+            vec![],
             vec![],
             vec![],
             #[cfg(feature = "orchard")]
@@ -5985,6 +5987,8 @@ pub(crate) fn query_nullifier_map<N: AsRef<[u8]>>(
             txid,
             index,
             vec![],
+            vec![],
+            #[cfg(feature = "transparent-inputs")]
             vec![],
             vec![],
             vec![],